server/src/requirements.txt
astroid==2.3.3
bcrypt==3.1.7
cachetools==4.1.0
cffi==1.14.0
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
google-auth==1.14.0
greenlet==0.4.15
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
python-engineio==3.11.2
python-socketio==4.4.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
